rename e2ee and update paths
