+++ modules/e2ee-internxt/SessionData.ts
